# Refactor hash calculation and token emission logic for improved clarity and performance